+++ packages/cloud/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chonkiejs/cloud
 
+## 0.0.5
+
+### Patch Changes
+
+- Fix: Add `embedding` to the `Chunk` for `EmbeddingsRefinery`
+- Updated dependencies
+  - @chonkiejs/core@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes