+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chonkiejs/core
 
+## 0.0.4
+
+### Patch Changes
+
+- Fix: Add `embedding` to the `Chunk` for `EmbeddingsRefinery`
+
 ## 0.0.3
 
 ### Patch Changes